Fix alpha and beta tubulin placement.
The placement of the child objects (AlphaUnit and BetaUnit) was offset
from the origin of AlphaVerts and BetaVerts. Technically, there's
nothing wrong with this, but it can make the transformation (i.e. scale,
translate, rotate) of AlphaVerts and BetaVerts harder to work with.

Additionally, AlphaVerts was not being renamed which prevented the
script from being run multiple times.